+++ backend/requirements.txt
@@ -1,4 +1,5 @@
 aiohttp==3.8.3
+asyncio
 aiosignal==1.3.1
 aiosmtplib==1.1.7
 alembic==1.8.1
